resolve cycle
