+++ deployables/extension/src/panel/utils/index.ts
@@ -1,4 +1,4 @@
-export { fetchContractInfo } from './abi'
+export { fetchContractInfo, type ContractInfo } from './abi'
 export { getStorageEntries } from './getStorageEntries'
 export { getStorageEntry } from './getStorageEntry'
 export { removeStorageEntry } from './removeStorageEntry'
